feat: progressive bit reduction per extra letter
This version stores only 24 bits for the blind index per part of the
name, so 24 for first and 24 for last.

It has 5 bits for the first letter, 4 for the second, 4 for the third, 4
for the fourth, 4 for the 5th, and 3 for the 6th. The hit rate seems
still pretty good on most, there are no false negatives due to the bloom
filter design, and if we don't limit the results too much, maybe skip
the 1 letter searches for being unreasonably useless, and reduce bits
later on with more bits for the early parts, we can get the performance
smoothed out. As it stands the accuracy approaches 100% around 4
letters, which is already past necessary, since it's good enough to
reduce the results greatly, no need to get 100% accurate results