feat: steeper bit reduction
1, 2, 3, 4, 5, and 6 letter bloom filter in a total of 24 bits per part
of name. So 48 bits per row. This gives us 6, 5, 4, 3, 3, 3 bits
respectively. We could even more aggressively expand this towards the
end, because the matches before make up for the inaccuracy of the
matches at the end. We could probably even combine 9 letters into this,
with 6, 5, 4, 3, 2, 1, 1, 1, 1 bits each, without expanding the space
form 24 bits